Add implementation to add new element dynamically
An already created tree will always have a base level
with a len that is a power of 2. So if we add a new element
this len changes. To get it to a power of 2 we do the same
as what we have been doing when creating a tree from an
input array that has a len different to a power of 2.
So when duplicating elements to reach the desired len,
we can create a new subtree from those new elements.
With these new subtree, the only thing we need to do is
concatenate and hash its root with the one in the merkle
tree, creating the new root and a new level on the tree.