Add support for adding when the tree has repeated elements
There are 2 cases to handle when adding an element to the tree.
First is the case when we add an element to a tree that already
has a base level that are all different elements and it has
a lenght that is power of 2. That case was already handled. The
other possible case is when the base level has repeated values.
Now this case is also handled by replacing the first repeated
value with the new element and re-calculating the part of the
tree affected by this change.